pyproject.toml
[tool.poetry]
name = "telebot_rzd"
version = "0.1.0"
description = ""
authors = ["Stefan <epiktetov.sv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
django-cleanup = "^8.0.0"
requests = "^2.31.0"
pytils = "^0.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^6.1.0"
flake8-django = "^1.4"
flake8-bugbear = "^23.9.16"
flake8-coding = "^1.3.2"
naming = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
